Delete item functionality implemented!
