+++ fashion-app/src/app/api/items/[id]/profile/route.ts
@@ -18,5 +18,20 @@ export async function GET(request:NextRequest, { params }: RouteParams) {
         console.log(error);
         return NextResponse.json({ message: "Error", error }, { status: 500 })
     }
+}
+
+export async function DELETE(request: NextRequest, {params}: RouteParams) {
+    const { id } = params;
+    if (!mongoose.Types.ObjectId.isValid(id)) {
+        return NextResponse.json({message: "Invalid ID format"}, {status: 400})
+    }
     
+    await connectMongoDB();
+    const deletedItem = await Item.findByIdAndDelete(id);
+    
+    if (!deletedItem) {
+        return NextResponse.json({ message: "Item not found"}, {status: 404});
+    }
+
+    return NextResponse.json({ message: "Item deleted" }, {status: 200});
 }
