+++ fashion-app/src/app/components/svg/DeleteSVG.tsx
@@ -1,7 +1,7 @@
 export default function DeleteSVG() {
     return (
         <svg width="31" height="31" viewBox="0 0 31 31" fill="none" xmlns="http://www.w3.org/2000/svg">
-        <circle cx="15.5" cy="15.5" r="15.5" fill="#FF2E16"/>
+        <circle cx="15.5" cy="15.5" r="15.5"/>
         <path d="M11.5151 22.6729C11.0767 22.6729 10.7015 22.5168 10.3893 22.2047C10.0772 21.8925 9.9211 21.5173 9.9211 21.0789V10.7181H9.12411V9.12413H13.109V8.32715H17.891V9.12413H21.8759V10.7181H21.0789V21.0789C21.0789 21.5173 20.9228 21.8925 20.6107 22.2047C20.2985 22.5168 19.9233 22.6729 19.4849 22.6729H11.5151ZM19.4849 10.7181H11.5151V21.0789H19.4849V10.7181ZM13.109 19.485H14.703V12.3121H13.109V19.485ZM16.297 19.485H17.891V12.3121H16.297V19.485Z" fill="white"/>
         </svg>
     )
