+++ fashion-app/src/app/components/svg/EditSVG.tsx
@@ -1,7 +1,7 @@
 export default function EditSVG() {
     return (
         <svg width="31" height="31" viewBox="0 0 31 31" fill="none" xmlns="http://www.w3.org/2000/svg">
-        <circle cx="15.5" cy="15.5" r="15.5" fill="#FFA216"/>
+        <circle cx="15.5" cy="15.5" r="15.5"/>
         <path d="M9.92109 21.0789H11.0568L18.8473 13.2884L17.7116 12.1527L9.92109 19.9432V21.0789ZM8.32712 22.6729V19.2857L18.8473 8.78542C19.0067 8.6393 19.1827 8.52639 19.3753 8.4467C19.5679 8.367 19.7705 8.32715 19.983 8.32715C20.1956 8.32715 20.4015 8.367 20.6007 8.4467C20.7999 8.52639 20.9726 8.64594 21.1187 8.80534L22.2146 9.92112C22.374 10.0672 22.4902 10.2399 22.5633 10.4392C22.6363 10.6384 22.6729 10.8377 22.6729 11.0369C22.6729 11.2494 22.6363 11.452 22.5633 11.6446C22.4902 11.8372 22.374 12.0132 22.2146 12.1726L11.7143 22.6729H8.32712ZM18.2695 12.7305L17.7116 12.1527L18.8473 13.2884L18.2695 12.7305Z" fill="white"/>
         </svg>
     )
